Changes to write map data to Library Folder
We will now be writing mapping and category data to the user's Library directory in macOS and removing the need for customer's to specify their own configuration files. This will simplify the user flow.

Note: This will only support macOS for now and we are removing support for other operating systems